Add overlay for 2 search boxes
- Simulate the UH main campus
