+++ Coog Path/Features/Home/ViewModels/LocationManager.swift
@@ -1,0 +1,64 @@
+//
+//  LocationManager.swift
+//  Coog Path
+//
+//  Created by Minh Nguyen on 10/27/23.
+//
+
+import Foundation
+import _MapKit_SwiftUI
+
+final class LocationManager: NSObject, ObservableObject {
+    private let locationManager = CLLocationManager()
+    
+    @Published var region = MapCameraPosition.region(MKCoordinateRegion(
+        center: CLLocationCoordinate2D(latitude: 29.766083, longitude: -95.358810),
+        span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)))
+    
+    
+    @Published var position: MapCameraPosition = .camera(MapCamera(centerCoordinate: CLLocationCoordinate2D(latitude: 29.72001, longitude: -95.34207), distance: 1000))
+//    @Published var position: MapCameraPosition = .userLocation(fallback: .automatic)
+    
+    override init() {
+        super.init()
+        
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBest
+        setup()
+    }
+    
+    func setup() {
+        switch locationManager.authorizationStatus {
+                // If we are authorized then we request location just once, to center the map
+            case .authorizedWhenInUse:
+                locationManager.requestLocation()
+                // If we don´t, we request authorization
+            case .notDetermined:
+                locationManager.startUpdatingLocation()
+                locationManager.requestWhenInUseAuthorization()
+            default:
+                break
+        }
+    }
+}
+
+extension LocationManager: CLLocationManagerDelegate {
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        guard .authorizedWhenInUse == manager.authorizationStatus else { return }
+        locationManager.requestLocation()
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+        print("Something went wrong: \(error)")
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        locationManager.stopUpdatingLocation()
+        locations.last.map {
+            region = MapCameraPosition.region(MKCoordinateRegion(
+                center: $0.coordinate,
+                span: .init(latitudeDelta: 0.01, longitudeDelta: 0.01)
+            ))
+        }
+    }
+}
